Fix code bug when calling fontconfig API in Context initialization. Add workaround to draw fonts like 	Adobe "Source Han Sans" which provides incorrect ascender/descender.

+++ intern/blendint/include/BlendInt/Gui/Font.hpp
@@ -79,7 +79,7 @@ namespace BlendInt {
 		Font();
 		
 		Font (const FcChar8* family,
-				double size,
+				int size,
 				int weight = FC_WEIGHT_REGULAR,
 				int slant = FC_SLANT_ROMAN);
 
@@ -120,7 +120,7 @@ namespace BlendInt {
 		 */
 		void SetWeight (int weight);
 
-		void SetSize (double size);
+		void SetSize (int size);
 
 		void SetPixelSize (double pixel_size);
